Create cy.createDashboardFromUnitTests function
Shared logic for creating dashboards for testing panels.
